chore: rename MainComposer to ContentEditor
