+++ src/components/ContentEditor/ContentEditor.stories.tsx
@@ -1,5 +1,5 @@
 import { Story } from '@ladle/react';
 
-import MainComposer from './MainComposer';
+import MainComposer from './ContentEditor';
 
 export const MainComposerStories: Story = () => <MainComposer />;
