+++ src/pages/home/home.tsx
@@ -3,9 +3,9 @@ import { ReactNode, useState } from 'react';
 import { useModule } from '~contexts/ModuleContext';
 
 import ComposerEditor from '~components/ComposerEditor/ComposerEditor';
+import MainComposer from '~components/ContentEditor/ContentEditor';
 import FeedbackError from '~components/FeedbackError/FeedbackError';
 import FirstComment from '~components/FirstComment/FirstComment';
-import MainComposer from '~components/MainComposer/MainComposer';
 import SavBar from '~components/SavBar/SavBar';
 
 import Header from './components/Header/Header';
